Create guest registration screen layout, and move view and viewModel files
