--- app/src/main/java/com/example/guest/GuestFromActivity.kt
@@ -1,11 +1,0 @@
-package com.example.guest
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class GuestFromActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_guest_from)
-    }
-}
