+++ app/src/main/java/com/example/guest/viewmodel/GuestFromViewModel.kt
@@ -1,0 +1,6 @@
+package com.example.guest.viewmodel
+
+import androidx.lifecycle.ViewModel
+
+class GuestFromViewModel : ViewModel() {
+}
