+++ app/src/main/java/com/example/guest/viewmodel/PresentViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.guest.ui.present
+package com.example.guest.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
