+++ app/src/main/res/layout/fragment_absent.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.absent.AbsentFragment">
+    tools:context=".view.AbsentFragment">
 
     <TextView
         android:id="@+id/text_absent"
