+++ app/src/main/res/layout/fragment_all_guests.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.allguestes.AllGuestsFragment">
+    tools:context=".view.AllGuestsFragment">
 
     <TextView
         android:id="@+id/text_all_guests"
