+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/nav_all"
-        android:name="com.example.guest.ui.allguestes.AllGuestsFragment"
+        android:name="com.example.guest.view.AllGuestsFragment"
         android:label="@string/menu_all_guests"
         tools:layout="@layout/fragment_all_guests" />
 
     <fragment
         android:id="@+id/nav_presents"
-        android:name="com.example.guest.ui.present.PresentFragment"
+        android:name="com.example.guest.view.PresentFragment"
         android:label="@string/menu_presents"
         tools:layout="@layout/fragment_present" />
 
     <fragment
         android:id="@+id/nav_absents"
-        android:name="com.example.guest.ui.absent.AbsentFragment"
+        android:name="com.example.guest.view.AbsentFragment"
         android:label="@string/menu_absents"
         tools:layout="@layout/fragment_absent" />
 </navigation>
